CH: Splitting up components into seperate files
