+++ frontend/src/components/Groups/GroupCapacity.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Group } from "../firebase/database";
+import { Group } from "../../firebase/database";
 import { Text } from "@chakra-ui/react";
 
 const GroupCapacity = (props: { group: Group }) => {
