+++ frontend/src/components/Header.tsx
@@ -23,7 +23,7 @@ import { ColorModeSwitcher } from "./ColorModeSwitcher";
 import { logout, auth } from "../firebase/firebase";
 import { MdEmail } from "react-icons/all";
 import { User } from "firebase/auth";
-import AddCar from "./AddCar";
+import AddCar from "./Profiles/AddCar";
 import { lightTheme } from "../theme/colours";
 
 export interface PageList {
