+++ frontend/src/components/Rides/ChooseCar.tsx
@@ -1,9 +1,9 @@
 import { Menu, Select, Spinner } from "@chakra-ui/react";
 import * as React from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { useUserVehicles, Vehicle } from "../firebase/database";
-import { auth } from "../firebase/firebase";
-import AddCar from "./AddCar";
+import { useUserVehicles, Vehicle } from "../../firebase/database";
+import { auth } from "../../firebase/firebase";
+import AddCar from "../Profiles/AddCar";
 
 const ChooseCar = (props: {
   carUpdate: (car: Vehicle) => void;
