+++ frontend/src/components/Rides/MapView.tsx
@@ -1,8 +1,8 @@
 import React, { CSSProperties, ReactNode } from "react";
 import { icon, LatLng, latLng, Map } from "leaflet";
 import { MapContainer, TileLayer } from "react-leaflet";
-import startIconImg from "../images/Arrow Circle Up_8.png";
-import endIconImg from "../images/Arrow Circle Down_8.png";
+import startIconImg from "../../images/Arrow Circle Up_8.png";
+import endIconImg from "../../images/Arrow Circle Down_8.png";
 
 const MAPBOX_ACCESS_TOKEN =
   "pk.eyJ1IjoibWFyY3VzZHVubiIsImEiOiJja3ppeTllOTAxanBuMm9uMnRwMHZ1dmF6In0.gHleMGVyUBmw_na8Elfzdg";
